Migrate tests into sub-directories
