+++ tests/data/color_map_test.py
