+++ tests/groups/remove_root_test.py
