+++ tests/objects/curve_data_test.py
