+++ tests/objects/geo_image_test.py
