+++ tests/objects/grid_object_test.py
