+++ tests/objects/maxwell_plate_test.py
