+++ tests/objects/text_object_test.py
